add :disable_cpt for resumable_upload/download

+++ lib/aliyun/oss/download.rb
@@ -13,7 +13,7 @@ module Aliyun
       class Download < Transaction
         def initialize(opts)
           super(opts)
-          @file, @checkpoint_file = opts[:file], opts[:resume_token]
+          @file, @checkpoint_file = opts[:file], opts[:cpt_file]
           @object_meta = {}
           @parts = []
         end
@@ -61,7 +61,7 @@ module Aliyun
         #     :md5 => 'states_md5'
         #   }
         def checkpoint!
-          logger.debug("Begin make checkpoint")
+          logger.debug("Begin make checkpoint, disable_cpt: #{options[:disable_cpt]}")
 
           ensure_object_not_changed
 
@@ -71,7 +71,8 @@ module Aliyun
             :object_meta => @object_meta,
             :parts => @parts
           }
-          write_checkpoint(states, @checkpoint_file)
+
+          write_checkpoint(states, @checkpoint_file) unless options[:disable_cpt]
 
           logger.debug("Done make checkpoint, states: #{states}")
         end
@@ -91,17 +92,17 @@ module Aliyun
             end
           end
 
-          File.delete(@checkpoint_file)
+          File.delete(@checkpoint_file) unless options[:disable_cpt]
           @parts.each{ |p| File.delete(get_part_file(p[:number])) }
 
           logger.info("Done commit transaction, id: #{id}")
         end
 
-        # Rebuild the states of the transaction from token file
+        # Rebuild the states of the transaction from checkpoint file
         def rebuild!
           logger.info("Begin rebuild transaction, checkpoint: #{@checkpoint_file}")
 
-          if File.exists?(@checkpoint_file)
+          if File.exists?(@checkpoint_file) and not options[:disable_cpt]
             states = load_checkpoint(@checkpoint_file)
 
             states[:parts].select{ |p| p[:done] }.each do |p|
